requirements.txt
Flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
openai
requests
sendgrid
gunicorn 
